Use cached discovered records on temporary discovery failure
When client run in endpoint discovery mode (-discover flag), it caches
SRV records lookup result on success; if in next time when client starts
with the same discover argument it temporarily fails to get SRV records
(i.e. network is down), it will use previously cached results instead of
failing.

This makes case or running client as a service on a laptop that change
networks often or may be offline more reliable.